requirements.txt
pandas
seaborn
scikit-learn
datasets
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git
transformers
peft
names
python-dotenv